// generated/site/orpc.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

import { oc } from '@orpc/contract';
import * as z from 'zod';

import { zSiteContactsCreateBody, zSiteContactsCreateResponse, zSiteContactsDeletePath, zSiteContactsGetPath, zSiteContactsGetResponse, zSiteContactsListQuery, zSiteContactsListResponse, zSiteContactsUpdateBody, zSiteContactsUpdatePath, zSiteContactsUpdateResponse } from './zod.gen.ts';

/**
 * List contacts for the site UI
 */
export const siteContactsList = oc.route({
    description: 'List contacts for the site UI',
    inputStructure: 'detailed',
    method: 'GET',
    operationId: 'SiteContacts_list',
    path: '/contacts',
    tags: ['Site Contacts']
}).input(z.object({ query: zSiteContactsListQuery.optional() })).output(zSiteContactsListResponse);

/**
 * Create a contact from the site UI
 */
export const siteContactsCreate = oc.route({
    description: 'Create a contact from the site UI',
    inputStructure: 'detailed',
    method: 'POST',
    operationId: 'SiteContacts_create',
    path: '/contacts',
    successStatus: 201,
    tags: ['Site Contacts']
}).input(z.object({ body: zSiteContactsCreateBody })).output(zSiteContactsCreateResponse);

/**
 * Delete a contact from the site UI
 */
export const siteContactsDelete = oc.route({
    description: 'Delete a contact from the site UI',
    inputStructure: 'detailed',
    method: 'DELETE',
    operationId: 'SiteContacts_delete',
    path: '/contacts/{id}',
    tags: ['Site Contacts']
}).input(z.object({ params: zSiteContactsDeletePath }));

/**
 * Get a contact by ID for the site UI
 */
export const siteContactsGet = oc.route({
    description: 'Get a contact by ID for the site UI',
    inputStructure: 'detailed',
    method: 'GET',
    operationId: 'SiteContacts_get',
    path: '/contacts/{id}',
    tags: ['Site Contacts']
}).input(z.object({ params: zSiteContactsGetPath })).output(zSiteContactsGetResponse);

/**
 * Update a contact from the site UI
 */
export const siteContactsUpdate = oc.route({
    description: 'Update a contact from the site UI',
    inputStructure: 'detailed',
    method: 'PUT',
    operationId: 'SiteContacts_update',
    path: '/contacts/{id}',
    tags: ['Site Contacts']
}).input(z.object({ body: zSiteContactsUpdateBody, params: zSiteContactsUpdatePath })).output(zSiteContactsUpdateResponse);

export const contract = {
    siteContactsList,
    siteContactsCreate,
    siteContactsDelete,
    siteContactsGet,
    siteContactsUpdate
};